+++ src/op/add_weighted_edge.rs
@@ -118,6 +118,42 @@ where
     }
 }
 
+// Slice
+
+impl<W> AddWeightedEdge<W> for &mut [Vec<(usize, W)>] {
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        self[s].push((t, w));
+    }
+}
+
+impl<W, H> AddWeightedEdge<W> for &mut [HashSet<(usize, W), H>]
+where
+    H: BuildHasher,
+    W: Eq + Hash,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert((t, w));
+    }
+}
+
+impl<W, H> AddWeightedEdge<W> for &mut [HashMap<usize, W, H>]
+where
+    H: BuildHasher,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert(t, w);
+    }
+}
+
 // Arr
 
 impl<const V: usize, W> AddWeightedEdge<W> for [Vec<(usize, W)>; V] {
@@ -332,6 +368,150 @@ mod tests {
                 [(1, 2), (2, 1)].iter().copied().collect(),
                 [(2, 4), (0, -2)].iter().copied().collect(),
                 once((0, 3)).collect()
+            ]
+        );
+    }
+
+    #[test]
+    fn slice_vec() {
+        let mut graph: [Vec<(usize, i32)>; 3] = [Vec::new(), Vec::new(), Vec::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(graph, [vec![(1, 2)], Vec::new(), Vec::new()]);
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(graph, [vec![(1, 2), (2, 1)], Vec::new(), Vec::new()]);
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(graph, [vec![(1, 2), (2, 1)], vec![(2, 4)], Vec::new()]);
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            [vec![(1, 2), (2, 1)], vec![(2, 4), (0, -2)], Vec::new()]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            [vec![(1, 2), (2, 1)], vec![(2, 4), (0, -2)], vec![(0, 3)]]
+        );
+    }
+
+    #[test]
+    fn slice_hash_set() {
+        let mut graph: [HashSet<(usize, i32)>; 3] =
+            [HashSet::new(), HashSet::new(), HashSet::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            [HashSet::from([(1, 2)]), HashSet::new(), HashSet::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            [
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::new(),
+                HashSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            [
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::from([(2, 4)]),
+                HashSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            [
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::from([(2, 4), (0, -2)]),
+                HashSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            [
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::from([(2, 4), (0, -2)]),
+                HashSet::from([(0, 3)])
+            ]
+        );
+    }
+
+    #[test]
+    fn slice_hash_map() {
+        let mut graph: [HashMap<usize, i32>; 3] = [HashMap::new(), HashMap::new(), HashMap::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            [HashMap::from([(1, 2)]), HashMap::new(), HashMap::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            [
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::new(),
+                HashMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            [
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::from([(2, 4)]),
+                HashMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            [
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::from([(2, 4), (0, -2)]),
+                HashMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            [
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::from([(2, 4), (0, -2)]),
+                HashMap::from([(0, 3)])
             ]
         );
     }